+++ x/evm/types/key.go
@@ -30,9 +30,9 @@ const (
 	// The EVM module should use a prefix store.
 	StoreKey = ModuleName
 
-	// TransientKey is the key to access the EVM transient store, that is reset
+	// ObjectStoreKey is the key to access the EVM object store, that is reset
 	// during the Commit phase.
-	TransientKey = "transient_" + ModuleName
+	ObjectStoreKey = "object:" + ModuleName
 
 	// RouterKey uses module name for routing
 	RouterKey = ModuleName
@@ -45,10 +45,11 @@ const (
 	prefixParams
 )
 
-// prefix bytes for the EVM transient store
+// prefix bytes for the EVM object store
 const (
-	prefixTransientBloom = iota + 1
-	prefixTransientGasUsed
+	prefixObjectBloom = iota + 1
+	prefixObjectGasUsed
+	prefixObjectParams
 )
 
 // KVStore key prefixes
@@ -58,10 +59,11 @@ var (
 	KeyPrefixParams  = []byte{prefixParams}
 )
 
-// Transient Store key prefixes
+// Object Store key prefixes
 var (
-	KeyPrefixTransientBloom   = []byte{prefixTransientBloom}
-	KeyPrefixTransientGasUsed = []byte{prefixTransientGasUsed}
+	KeyPrefixObjectBloom   = []byte{prefixObjectBloom}
+	KeyPrefixObjectGasUsed = []byte{prefixObjectGasUsed}
+	KeyPrefixObjectParams  = []byte{prefixObjectParams}
 )
 
 // AddressStoragePrefix returns a prefix to iterate over a given account storage.
@@ -74,16 +76,16 @@ func StateKey(address common.Address, key []byte) []byte {
 	return append(AddressStoragePrefix(address), key...)
 }
 
-func TransientGasUsedKey(txIndex int) []byte {
-	var key [9]byte
-	key[0] = prefixTransientGasUsed
+func ObjectGasUsedKey(txIndex int) []byte {
+	var key [1 + 8]byte
+	key[0] = prefixObjectGasUsed
 	binary.BigEndian.PutUint64(key[1:], uint64(txIndex))
 	return key[:]
 }
 
-func TransientBloomKey(txIndex, msgIndex int) []byte {
+func ObjectBloomKey(txIndex, msgIndex int) []byte {
 	var key [1 + 8 + 8]byte
-	key[0] = prefixTransientBloom
+	key[0] = prefixObjectBloom
 	binary.BigEndian.PutUint64(key[1:], uint64(txIndex))
 	binary.BigEndian.PutUint64(key[9:], uint64(msgIndex))
 	return key[:]